refactor: split frame and dimensions into component and hooks respectively
