+++ src/components/Mirror.tsx
@@ -1,8 +1,7 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import { getBounds, isElement } from "../utils/dom";
 import { randomString } from "../utils";
-import { useObserver } from "../hooks/useObserver";
+import { useDimensions } from "../hooks/useDimensions";
 import { Frame, FrameProps } from "./Frame";
 import { Reflection } from "./Reflection";
 
@@ -25,27 +24,12 @@ export function Mirror({ className, reflect }: MirrorProps) {
         }),
         [reflect],
     );
-    const dimens = useDimensions(real);
+
+    const { height, width } = useDimensions(real);
+    const frameProps = { className, height, width, id: instanceId };
     return (
-        <Frame
-            className={className}
-            height={dimens.height}
-            id={instanceId}
-            width={dimens.width}
-        >
+        <Frame {...frameProps}>
             <Reflection real={real} style={{ pointerEvents: "none" }} />
         </Frame>
     );
-}
-
-function useDimensions(node?: Node) {
-    const [dimensions, setDimensions] = React.useState(new DOMRect());
-    const target = React.useMemo(() => {
-        return isElement(node) ? node : undefined;
-    }, [node]);
-    const onUpdate = React.useCallback(() => {
-        setDimensions(getBounds(target));
-    }, [target]);
-    useObserver({ ObserverClass: ResizeObserver, onUpdate, target });
-    return dimensions;
 }
